playwright/llm-automation/requirements.txt
# Core dependencies
streamlit>=1.28.0
playwright>=1.40.0
openai>=1.3.0
python-dotenv>=1.0.0

# LangChain and AI
altair>=5.0.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-core>=0.1.0
langchain-openai>=0.0.1
langchain-experimental>=0.0.1

# Additional utilities
asyncio-mqtt>=0.11.1
aiofiles>=23.2.1
Pillow>=10.0.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
# types-python-dotenv not available for Python 3.13 yet

# Testing and debugging
pytest-playwright>=0.4.0
pytest-html>=4.0.0
pytest-cov>=4.0.0